self_rag/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface-hub>=0.15.0
scikit-learn>=1.2.0
safetensors>=0.3.0
bitsandbytes>=0.40.0
triton>=2.0.0
einops>=0.6.1
beir>=1.0.0
sentence-transformers>=2.2.2
pandas>=1.5.0
nltk>=3.8.1
scipy>=1.10.0
faiss-cpu>=1.7.4
vllm>=0.2.0
backoff>=2.2.1
jsonlines>=3.1.0
openai>=1.0.0
spacy>=3.5.0
